+++ examples/probabilistic/slam2/sir.h
@@ -179,7 +179,7 @@ template <typename PTYPE, typename BASE> struct SIR_t : public PF_t<PTYPE, BASE>
 
     virtual std::string id() const {
         std::string id_str;
-        id_str = std::string("PF(type=SIR,nparticles=") + std::to_string(nparticles_)
+        id_str = std::string("PF(type=SIR,nparticles=") + std::to_string((long long)nparticles_)
           //+ ",sus=" + std::to_string(do_stochastic_universal_sampling_)
           + ")";
         return id_str;
